Rename `prefect_aws.cli` to `prefect_aws._cli` and remove `CLI_USAGE.md`
